cockpit.js: split Channel into its own file
Another straight move.

This makes it possible to create channels without needing to import
`cockpit.js`.
